Revert "feat(ai-agent): support graphql subscriptions (#2448)"
This reverts commit 1fdf60dd90a282f5c291139f3ea7e3c400cf84f6.
