+++ package.json
@@ -115,16 +115,13 @@
     "@nangohq/frontend": "^0.67.8",
     "@sentry/react": "9.11.0",
     "@tanstack/react-virtual": "3.11.2",
-    "@types/actioncable": "^5.2.11",
     "ace-builds": "^1.37.3",
-    "actioncable": "5.2.8-1",
     "apollo-link-timeout": "4.0.0",
     "apollo-upload-client": "17.0.0",
     "apollo3-cache-persist": "0.15.0",
     "decimal.js": "^10.4.3",
     "formik": "2.4.6",
     "graphql": "^16.6.0",
-    "graphql-ruby-client": "^1.14.8",
     "html-react-parser": "^5.2.6",
     "lago-configs": "workspace:*",
     "lago-design-system": "workspace:*",
